images-descriptor/requirements.txt
Pillow==9.4.0
boto3==1.26.79
pynamodb==5.4.1
requests==2.28.2
